+++ fed_kit_client/ios/Runner/Utils/MLHelper.swift
@@ -70,7 +70,7 @@ struct Layer {
 }
 
 /// Assuming the two have the same length.
-func meanSquareErrors(_ a: [Float], _ b: [Int32]) -> Float {
+func meanSquareErrors(_ a: [Float], _ b: [Double]) -> Float {
     var sum: Float = 0.0
     for (index, value) in a.enumerated() {
         let diff = value - Float(b[index])
